Docs: Quick README fix for clarity
I was doing a workshop, and I needed a way to show a commit that was
maybe a little bit contrived but it's okay, realy all that matters is
there is a long form commit message happening here. Please see issue #1,
there is nothing in there, but I really wanted to reference issue #1
because it's cool to see what happens. Ping @shiffman too because he
made me do this